# server/user-service/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: UserService API
    version: 0.0.1
paths:
    /v1/user:
        get:
            tags:
                - UserService
            operationId: UserService_ListUser
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListUserReply'
        post:
            tags:
                - UserService
            operationId: UserService_CreateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateUserReply'
    /v1/user/{id}:
        get:
            tags:
                - UserService
            operationId: UserService_GetUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetUserReply'
        put:
            tags:
                - UserService
            operationId: UserService_UpdateUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateUserReply'
        delete:
            tags:
                - UserService
            operationId: UserService_DeleteUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteUserReply'
components:
    schemas:
        CreateUserReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
        CreateUserRequest:
            type: object
            properties:
                groupId:
                    type: integer
                    format: int32
                username:
                    type: string
                password:
                    type: string
                mobile:
                    type: string
                email:
                    type: string
        DeleteUserReply:
            type: object
            properties: {}
        GetUserReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/CreateUserRequest'
        ListUserReply:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/CreateUserRequest'
        UpdateUserReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
        UpdateUserRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                data:
                    $ref: '#/components/schemas/CreateUserRequest'
tags:
    - name: UserService